02_AgenticAIFunction_Login/requirements.txt
azure-functions
azure-identity
azure-ai-projects
azure-ai-agents
requests
